Added utility functions & fix for missing zero
This change adds a fix to prepend a zero in the hour place of IGT
and RTA times if it is missing. The missing zero was causing sorting
problems when displaying the leaderboard.

It also adds a couple utility functions. These commands are race
creator only and hidden from the help:
    $user_races - Will show race submissions for the provided user_id
    $edit_submission - Edits the IGT, CR and RTA for any async submission
    $mod_util - Adds support for several simple operations: force update
        leaderboard, force update weekly submit channel message, toggle
        toilet paper reaction on/off.